requirements.txt
certifi==2024.8.30
chardet==4.0.0
colorama==0.4.6
configparser==7.1.0
crayons==0.4.0
idna==2.10
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.25.1
selenium==3.141.0
six==1.17.0
tzdata==2024.2
urllib3==1.26.20
webdriver-manager==3.4.2
